+++ pkgs/unified_analytics/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 8.0.5
+- Fix `Event.flutterWasmDryRun` fields.
+
+## 8.0.4
+- Changed `Event.flutterWasmDryRun` to track dart2wasm dry run metrics from Flutter.
+
 ## 8.0.3
 - Changed `Event.contextStructure` to make optional data that's no longer being
   collected and to add data about the size of library cycles.